Change dependencies to API and improve type consistency
Switched `kotlinx.browser` dependency from `implementation` to `api` for better reusability. Updated types to `UInt` and standardized naming conventions for improved code clarity and consistency.

+++ wgpu4k-scenes/src/jsMain/kotlin/io.ygdrasil.wgpu.examples/AssetManager.js.kt
@@ -11,5 +11,5 @@ actual var customVfs: Vfs
     set(value) {}
 
 internal actual fun Bitmap32.toBitmapHolder(textureFormat: ColorFormat): ImageBitmapHolder {
-    return ImageBitmapHolder(width, height, extractBytes(textureFormat))
+    return ImageBitmapHolder(width.toUInt(), height.toUInt(), extractBytes(textureFormat))
 }
